# crates/chat-server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "scion-chat",
    "description": "A chat server over HTTP/3-over-SCION",
    "contact": {
      "name": "Anapaya & contributors"
    },
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/rooms": {
      "get": {
        "tags": [
          "rooms"
        ],
        "summary": "List every room.",
        "operationId": "list",
        "responses": {
          "200": {
            "description": "Every room on the server",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomsResponse"
                }
              }
            }
          },
          "401": {
            "description": "No usable bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "rooms"
        ],
        "summary": "Create a room, or return the one already holding the name.",
        "description": "Creation is idempotent on the name, so a client never has to look the id up separately.",
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoomRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The name was taken; this is the room holding it",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          },
          "201": {
            "description": "The room was created",
            "headers": {
              "location": {
                "schema": {
                  "type": "string"
                },
                "description": "Where the new room is"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          },
          "401": {
            "description": "No usable bearer token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "The name is not acceptable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "The server accepts no more rooms",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "description": "The contents of an [`ErrorResponse`].",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "What failed, as a value a client can branch on. The HTTP status alone is too coarse."
          },
          "message": {
            "type": "string",
            "description": "A human-readable explanation, for logs and for showing to a user. Free-form: it may change\nbetween server versions, so never branch on it."
          }
        }
      },
      "CreateRoomRequest": {
        "type": "object",
        "description": "The name a new room is created under.\n\nCreation is idempotent on the name: an existing name returns that room instead of failing.",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "The name to create. Printable ASCII, 1–64 characters; matched case-insensitively against\nthe rooms that already exist."
          }
        }
      },
      "ErrorCode": {
        "oneOf": [
          {
            "type": "string",
            "description": "The request body was missing, malformed, or not the shape the endpoint expects.",
            "enum": [
              "invalid_body"
            ]
          },
          {
            "type": "string",
            "description": "The server failed for a reason it does not describe further.",
            "enum": [
              "internal"
            ]
          },
          {
            "type": "string",
            "description": "The request carried no usable bearer token.",
            "enum": [
              "unauthorized"
            ]
          },
          {
            "type": "string",
            "description": "The username exists but the password does not match it.",
            "enum": [
              "invalid_credentials"
            ]
          },
          {
            "type": "string",
            "description": "The requested username is already registered.",
            "enum": [
              "username_taken"
            ]
          },
          {
            "type": "string",
            "description": "The requested username is outside what registration accepts.",
            "enum": [
              "invalid_username"
            ]
          },
          {
            "type": "string",
            "description": "The requested room name is outside what room creation accepts.",
            "enum": [
              "invalid_name"
            ]
          },
          {
            "type": "string",
            "description": "No room has the requested id.",
            "enum": [
              "room_not_found"
            ]
          },
          {
            "type": "string",
            "description": "The message body exceeds the server's limit.",
            "enum": [
              "message_too_large"
            ]
          },
          {
            "type": "string",
            "description": "A configured cap — accounts or rooms — is already reached.",
            "enum": [
              "cap_exceeded"
            ]
          },
          {
            "type": "object",
            "description": "A code this build does not know, kept verbatim.\n\nDecoding an error must not itself fail: a client that meets a code added after it was\nbuilt still surfaces the server's message rather than a parse error.",
            "required": [
              "other"
            ],
            "properties": {
              "other": {
                "$ref": "#/components/schemas/UnknownCode",
                "description": "A code this build does not know, kept verbatim.\n\nDecoding an error must not itself fail: a client that meets a code added after it was\nbuilt still surfaces the server's message rather than a parse error."
              }
            }
          }
        ],
        "description": "Every failure this API distinguishes, on the wire as its name in `snake_case`.\n\nAn enum rather than a string so the server cannot emit a code that does not exist and a client\ncan `match` with the compiler checking it has covered them all."
      },
      "ErrorResponse": {
        "type": "object",
        "description": "The body of every failing response, whatever failed.",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError",
            "description": "What went wrong."
          }
        }
      },
...
      "Room": {
        "type": "object",
        "description": "A room.\n\nEvery user belongs to every room: there is no membership, no joining, and no private rooms.\nA room named `lobby` always exists.",
        "required": [
          "id",
          "name",
          "latest_seq"
        ],
        "properties": {
          "id": {
            "$ref": "#/components/schemas/RoomId",
            "description": "The room's identifier, used in the message paths."
          },
          "latest_seq": {
            "$ref": "#/components/schemas/Seq",
            "description": "The `seq` of the newest message in the room, or `0` while it has none.\n\nComparing it against the newest `seq` a client has read is what drives an unread badge, so\none listing covers every room at once. Not a count: `seq` is server-wide (see [`Seq`])."
          },
          "name": {
            "type": "string",
            "description": "The room's display name. Printable ASCII, 1–64 characters; unique case-insensitively."
          }
        }
      },
      "RoomId": {
        "type": "integer",
        "format": "int64",
        "description": "Identifier of a room, assigned by the server when the room is created. Stable for the room's\nlifetime, and rooms are never deleted.",
        "minimum": 0
      },
      "RoomsResponse": {
        "type": "object",
        "description": "Every room on the server.",
        "required": [
          "rooms"
        ],
        "properties": {
          "rooms": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Room"
            },
            "description": "All rooms. Never empty: `lobby` is always present."
          }
        }
      },
      "Seq": {
        "type": "integer",
        "format": "int64",
        "description": "Position of a message in the server's message sequence.\n\n`seq` increases strictly and never regresses, so it is the one total order that every client\nagrees on. It is assigned server-wide rather than per room, so a single room's messages carry\ngaps: treat `seq` as a cursor, never as a count. Clients remember the highest `seq` they have\nseen in a room and poll for what came after it.\n\nNumbering starts at 1, so [`Seq::START`] is the position before every message — an empty room\nand a client that has read nothing both sit there, and asking for what follows it asks for\neverything.",
        "minimum": 0
      },
...
      "UnknownCode": {
        "type": "string",
        "description": "A code carried by a response but not named by [`ErrorCode`].\n\nIts contents are private and it has no constructor, so decoding is the only thing that can\nproduce one. That is what makes a sender structurally unable to invent a code: reaching for\n[`ErrorCode::Other`] instead of adding a variant is not an option it has."
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}